+++ test_jbxapi.py
@@ -73,29 +73,6 @@ def test_file_submission_tuple(joe, monkeypatch):
     assert mock.requests[0].files["sample"] == ("Filename", sample)
 
 
-def test_strange_file_names(joe, monkeypatch):
-    names = {
-        "Sample": "Sample",
-        "\xc3\xb6": "xc3xb6",
-        "|": "x7c",
-    }
-
-    mock = MockedResponse(ok=True, json=successful_submission)
-    monkeypatch.setattr("requests.sessions.Session.post", mock)
-
-    for i, (name, expected) in enumerate(names.items()):
-        s = io.BytesIO(b"Testdata")
-        s.name = name
-        joe.submit_sample(s, cookbook=s)
-        assert mock.requests[i * 2].files["sample"] == (expected, s)
-        assert mock.requests[i * 2].files["cookbook"] == (expected, s)
-
-        s = io.BytesIO(b"Testdata")
-        joe.submit_sample((name, s), cookbook=(name, s))
-        assert mock.requests[i * 2 + 1].files["sample"] == (expected, s)
-        assert mock.requests[i * 2 + 1].files["cookbook"] == (expected, s)
-
-
 def test_url_submission(joe, monkeypatch):
     mock = MockedResponse(ok=True, json=successful_submission)
     monkeypatch.setattr("requests.sessions.Session.post", mock)
@@ -176,6 +153,17 @@ def test_array_parameters_single_value(joe, monkeypatch):
 
     assert mock.requests[0].data["tags[]"] == "mytag"
     assert "tags" not in mock.requests[0].data
+
+
+def test_user_agent():
+    joe = jbxapi.JoeSandbox()
+    assert "jbxapi.py" in joe.session.headers["User-Agent"]
+    assert jbxapi.__version__ in joe.session.headers["User-Agent"]
+
+    joe = jbxapi.JoeSandbox(user_agent="My Integration")
+    assert "jbxapi.py" in joe.session.headers["User-Agent"]
+    assert "My Integration" in joe.session.headers["User-Agent"]
+    assert jbxapi.__version__ in joe.session.headers["User-Agent"]
 
 
 # CLI tests